+++ src/yuv/gbrapf32.rs
@@ -73,10 +73,10 @@ pub fn gbrapf32_to<S: Gbrapf32Sink>(src: &Gbrapf32Frame<'_>, sink: &mut S) -> Re
   let b_plane = src.b();
   let r_plane = src.r();
   let a_plane = src.a();
-  let g_stride = src.g_stride();
-  let b_stride = src.b_stride();
-  let r_stride = src.r_stride();
-  let a_stride = src.a_stride();
+  let g_stride = src.g_stride() as usize;
+  let b_stride = src.b_stride() as usize;
+  let r_stride = src.r_stride() as usize;
+  let a_stride = src.a_stride() as usize;
 
   for row in 0..h {
     let g = &g_plane[row * g_stride..row * g_stride + w];
