+++ src/yuv/gbrpf32.rs
@@ -65,9 +65,9 @@ pub fn gbrpf32_to<S: Gbrpf32Sink>(src: &Gbrpf32Frame<'_>, sink: &mut S) -> Resul
   let g_plane = src.g();
   let b_plane = src.b();
   let r_plane = src.r();
-  let g_stride = src.g_stride();
-  let b_stride = src.b_stride();
-  let r_stride = src.r_stride();
+  let g_stride = src.g_stride() as usize;
+  let b_stride = src.b_stride() as usize;
+  let r_stride = src.r_stride() as usize;
 
   for row in 0..h {
     let g = &g_plane[row * g_stride..row * g_stride + w];
